Add unit test of event minor version bump
