--- source/dotnet/wholesale-integration-events-package/Contracts.Tests/CalculationResultCompleted.cs
@@ -1,45 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using Energinet.DataHub.Wholesale.Contracts.Events;
-using FluentAssertions;
-using Xunit;
-
-namespace Contracts.Tests;
-
-public class CalculationResultCompletedTests
-{
-    [Fact]
-    public void CalculationResultCompleted_HasCorrectNamespace()
-    {
-        // Be careful to change the namespace as it is public available
-        typeof(CalculationResultCompleted).Namespace.Should().Be("Energinet.DataHub.Wholesale.Contracts.Events");
-    }
-
-    [Fact]
-    public void CalculationResultCompleted_CodeIsGenerated()
-    {
-        // Assert that the code has been generated by an heuristic - if the CalculationResultCompleted.Parser property exists then the type has most likely been generated
-        // If not, the test code won't even compile. Note that it is not enough to test for existence of the type as we've created a custom partial class.
-        // Actually the ".Should().NotBeNull()" part is not necessary, but it's required by the compiler.
-        CalculationResultCompleted.Parser.Should().NotBeNull();
-    }
-
-    [Fact]
-    public void CalculationResultCompleted_HasCorrectEventName()
-    {
-        // Be careful to change the event name as it is public available and used for transport message meta data in accordance with ADR-008
-        CalculationResultCompleted.EventName.Should().Be("CalculationResultCompleted");
-    }
-}
