+++ source/dotnet/wholesale-integration-events-package/Contracts.Tests/CalculationResultCompletedTests.cs
@@ -1,23 +1,89 @@
-/* Copyright 2020 Energinet DataHub A/S
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Energinet.DataHub.Wholesale.Contracts.Events;
+using FluentAssertions;
+using Xunit;
+
+namespace Contracts.Tests;
+
+public class CalculationResultCompletedTests
+{
+    [Fact]
+    public void CalculationResultCompleted_HasCorrectNamespace()
+    {
+        // Be careful to change the namespace as it is public available
+        typeof(CalculationResultCompleted).Namespace.Should().Be("Energinet.DataHub.Wholesale.Contracts.Events");
+    }
+
+    [Fact]
+    public void CalculationResultCompleted_CodeIsGenerated()
+    {
+        // Assert that the code has been generated by an heuristic - if the CalculationResultCompleted.Parser property exists then the type has most likely been generated
+        // If not, the test code won't even compile. Note that it is not enough to test for existence of the type as we've created a custom partial class.
+        // Actually the ".Should().NotBeNull()" part is not necessary, but it's required by the compiler.
+        CalculationResultCompleted.Parser.Should().NotBeNull();
+    }
+
+    [Fact]
+    public void CalculationResultCompleted_HasCorrectEventName()
+    {
+        // Be careful to change the event name as it is public available and used for transport message meta data in accordance with ADR-008
+        CalculationResultCompleted.EventName.Should().Be("CalculationResultCompleted");
+    }
+
+    /// <summary>
+    /// If this test fails, it probably means that the minor event version has not been bumped, but the contract has changed.
+    /// In order to fix you should (probably):
+    /// - Bump the minor event version
+    /// - Update the content of the contract reference <see cref="LastKnownContentOfContract"/>
+    /// </summary>
+    [Fact]
+    public void Fail_If_MinorEventVersion_ShouldHaveBeenBumped()
+    {
+        // Arrange
+        var actualVersion = CalculationResultCompleted.EventMinorVersion;
+        var actualContent = File.ReadAllText(@"calculation_result_completed.proto");
+
+        // Act: There is no assert but the test will fail if the minor event version has not been bumped while the contract has changed.
+
+        // Assert (using Assert as it provides a more useful diff than FluentAssertions)
+        actualVersion.Should().Be(LastKnownMinorEventVersion);
+        Assert.Equal(LastKnownContentOfContract, actualContent, ignoreLineEndingDifferences: true, ignoreWhiteSpaceDifferences: true);
+    }
+
+    private const int LastKnownMinorEventVersion = 2;
+    private const string LastKnownContentOfContract = @"/* Copyright 2020 Energinet DataHub A/S
  *
- * Licensed under the Apache License, Version 2.0 (the "License2");
+ * Licensed under the Apache License, Version 2.0 (the ""License2"");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
+ * distributed under the License is distributed on an ""AS IS"" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
 
-syntax = "proto3";
-import "google/protobuf/timestamp.proto";
-import "google/protobuf/wrappers.proto";
+syntax = ""proto3"";
+import ""google/protobuf/timestamp.proto"";
+import ""google/protobuf/wrappers.proto"";
 
-option csharp_namespace = "Energinet.DataHub.Wholesale.Contracts.Events";
+option csharp_namespace = ""Energinet.DataHub.Wholesale.Contracts.Events"";
 
 /*
  * CalculationResultReady Integration Event.
@@ -224,4 +290,6 @@ message DecimalValue {
   // Nano units of the amount (10^-9)
   // Must be same sign as units
   sfixed32 nanos = 2;
+}
+";
 }
