+++ source/dotnet/wholesale-integration-events-package/Contracts.Tests/Contracts.Tests.csproj
@@ -22,4 +22,11 @@
       <ProjectReference Include="..\Contracts\Contracts.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Content Include="..\..\wholesale-api\Events\Events.Infrastructure\IntegrationEvents\calculation_result_completed.proto">
+        <Link>calculation_result_completed.proto</Link>
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </Content>
+    </ItemGroup>
+
 </Project>
